Update after move parser to new file
