+++ dnevnikru/exceptions.py
@@ -1,0 +1,4 @@
+class DnevnikError(Exception):
+    def __init__(self, message, errors):
+        super().__init__(message)
+        self.errors = f'DnevnikException[{errors}]'
